requirements.txt
requests
pandas
newspaper3k
textblob
boto3
python-dotenv